+++ packages/core/src/cache/cache.ts
@@ -16,6 +16,7 @@ import imageIdToURI from '../utilities/imageIdToURI';
 import eventTarget from '../eventTarget';
 import Events from '../enums/Events';
 import { ImageQualityStatus } from '../enums';
+import fnv1aHash from '../utilities/fnv1aHash';
 
 const ONE_GB = 1073741824;
 
@@ -33,12 +34,44 @@ class Cache {
   private readonly _imageCache = new Map<string, ICachedImage>();
   // used to store volume data (3d)
   private readonly _volumeCache = new Map<string, ICachedVolume>();
+  // used to store the reverse lookup from imageIds to volumeId
+  private readonly _imageIdsToVolumeIdCache = new Map<string, string>();
   // Todo: contour for now, but will be used for surface, etc.
   private readonly _geometryCache = new Map<string, ICachedGeometry>();
 
   private _imageCacheSize = 0;
   private _maxCacheSize = 3 * ONE_GB;
   private _geometryCacheSize = 0;
+
+  /**
+   * Generates a deterministic volume ID from a list of image IDs
+   * @param imageIds - Array of image IDs
+   * @returns A deterministic volume ID
+   */
+  public generateVolumeId(imageIds: string[]): string {
+    const imageURIs = imageIds.map(imageIdToURI).sort();
+
+    let combinedHash = 0x811c9dc5;
+    for (const id of imageURIs) {
+      const idHash = fnv1aHash(id);
+      for (let i = 0; i < idHash.length; i++) {
+        combinedHash ^= idHash.charCodeAt(i);
+        combinedHash +=
+          (combinedHash << 1) +
+          (combinedHash << 4) +
+          (combinedHash << 7) +
+          (combinedHash << 8) +
+          (combinedHash << 24);
+      }
+    }
+    return `volume-${(combinedHash >>> 0).toString(36)}`;
+  }
+
+  public getImageIdsForVolumeId(volumeId: string): string[] {
+    return Array.from(this._imageIdsToVolumeIdCache.entries())
+      .filter(([_, id]) => id === volumeId)
+      .map(([key]) => key);
+  }
 
   /**
    * Set the maximum cache Size